add fetch all products

+++ Abysalto.API/Models/Product.cs
@@ -4,7 +4,7 @@
     {
         int Id { get; set; }
         string Title { get; set; }
-        float Price { get; set; }
+        decimal Price { get; set; }
         string Category { get; set; }
         string Description { get; set; }
         string ImageUrl { get; set; }
